scripts: allow to install apps globally

+++ scripts/install/install.sh
@@ -7,18 +7,23 @@ url="$2"
 hash="$3"
 
 if [ -z "$name" ] || [ -z "$url" ] || [ -z "$hash" ]; then
-    echo "usage: $(basename $0) <name> <url> <hash>"
+    echo "usage: $(basename $0) <name>|global <url> <hash>"
     exit 1
 fi
 
-cd "$(dirname $0)/../.."
-
 file="${url##*/}"
-install_dir="apps/$name"
 
-rm -rf "$install_dir"
-mkdir -p "$install_dir"
-cd "$install_dir"
+if [ "$name" = "global" ]; then
+    cd /usr/local
+else
+    cd "$(dirname $0)/../.."
+
+    install_dir="apps/$name"
+
+    rm -rf "$install_dir"
+    mkdir -p "$install_dir"
+    cd "$install_dir"
+fi
 
 curl --silent --fail --show-error --location -o "$file" "$url"
 echo "$hash *$file" | sha256sum -c
